pyproject.toml
[project]
name = "in_netsuite"
version = "0.1.0"
description = "A Python package for connecting to NetSuite"
requires-python = ">=3.11"

dependencies = [
    "oauthlib>=3.2.2",
    "pydantic>=2.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "async-lru>=2.0.5",
    "black==25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest_asyncio",
    "black",
    "isort",
    "ruff",
    "pytailwindcss",
    "pip",
    "setuptools",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(levelname)s:%(name)s:%(message)s"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"

[tool.mypy]
files = "src/"
strict = true

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "C", "N", "B"]
ignore = [
    # Ignore specific rules if needed
]
target-version = "py38"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
    "venv",
    ".venv",
]

[tool.ruff.isort]
# Use isort compatibility within ruff
combine-as-imports = true
known-third-party = ["pillow", "numpy"]  # Add known third-party libraries
